Feature/update notifications

+++ src/app/pages/main/header/notifications-dropdown-menu/notifications-dropdown-menu.component.ts
@@ -1,12 +1,14 @@
 import { formatDate } from '@angular/common';
 import {
-  Component, ElementRef, HostListener, OnInit, Renderer2, ViewChild
+  Component, ElementRef, EventEmitter, HostListener, OnInit, Output, Renderer2, ViewChild
 } from '@angular/core';
 import { Router } from '@angular/router';
 import { DateUtils } from 'src/app/models/common.models';
 import { UserNotification } from 'src/app/models/notification.model';
+import { Suite } from 'src/app/models/suite.models';
 import { Logger, LoggerManager } from 'src/app/utils/logging';
 import { AppService } from 'src/app/utils/services/app.service';
+import { InputDialogService } from 'src/app/utils/services/input-dialog.service';
 
 @Component({
   selector: 'app-notifications-dropdown-menu',
@@ -14,14 +16,10 @@ import { AppService } from 'src/app/utils/services/app.service';
   styleUrls: ['./notifications-dropdown-menu.component.scss'],
 })
 export class NotificationsDropdownMenuComponent implements OnInit {
-  @ViewChild('dropdownMenu', { static: false }) dropdownMenu;
 
-  @HostListener('document:click', ['$event'])
-  clickout(event) {
-    if (!this.elementRef.nativeElement.contains(event.target)) {
-      this.hideDropdownMenu();
-    }
-  }
+  @ViewChild('dropdownMenu', { static: false }) dropdownMenu: any;
+
+  @Output() openUserProfile = new EventEmitter<boolean>();
 
   public notifications: UserNotification[];
   private notificationsByDate: {};
@@ -33,7 +31,8 @@ export class NotificationsDropdownMenuComponent implements OnInit {
     private router: Router,
     private elementRef: ElementRef,
     private renderer: Renderer2,
-    private appService: AppService) { }
+    private appService: AppService,
+    private inputDialog: InputDialogService) { }
 
   ngOnInit() {
     this.appService.loadNotifications().subscribe((data: UserNotification[]) => {
@@ -42,10 +41,17 @@ export class NotificationsDropdownMenuComponent implements OnInit {
     })
   }
 
+  @HostListener('document:click', ['$event'])
+  clickout(event: { target: any; }) {
+    if (!this.elementRef.nativeElement.contains(event.target)) {
+      this.hideDropdownMenu();
+    }
+  }
+
   private updateNotifications(notifications: UserNotification[]) {
     this.logger.debug("Updating notifications...", notifications);
-    this.notifications = notifications;
-    this.notificationsByDate = this.groupNotificationsByDate(notifications);
+    this.notifications = notifications.filter(n => n.event !== 'UNCONFIGURED_EMAIL' || !n.read);
+    this.notificationsByDate = this.groupNotificationsByDate(this.notifications);
     this.logger.debug("Notifications by date: ", this.notifications);
   }
 
@@ -74,13 +80,38 @@ export class NotificationsDropdownMenuComponent implements OnInit {
     return result;
   }
 
-  public setNotificationAsRead(n: UserNotification) {
+  public readNotification(n: UserNotification) {
     this.appService.setNotificationsReadingTime([n]).subscribe((data) => {
       this.logger.debug("Notification marked as read", n);
       this.hideDropdownMenu();
     });
+    if (n.event === 'UNCONFIGURED_EMAIL') {
+      // this.openUserProfile.emit(true);
+    } else if (n.event === 'BUILD_FAILED' || n.event === 'BUILD_RECOVERED') {
+      let suite: Suite = null;
+      if (n.data && "build" in n.data
+        && "suite" in n.data["build"]
+        && "workflow" in n.data["build"]) {
+        suite = this.appService.findTestSuite(
+          n.data["build"]["suite"]["uuid"],
+          n.data["build"]["workflow"]["uuid"]
+        );
+      }
+      this.logger.debug("Test suite related with notification", suite);
+      if (!suite || suite === undefined) {
+        this.inputDialog.show({
+          iconClass: 'fas fa-exclamation-triangle text-warning',
+          question: "Ops...",
+          description:
+            'Unable to find the test suite related with this notification',
+          confirmText: "",
+          cancelText: "Close",
+        });
+      } else {
+        return this.navigateTo('/suite', { 's': suite.asUrlParam() });
+      }
+    }
   }
-
 
   public markAllAsRead(notifications: UserNotification[] = null) {
     this.appService.setNotificationsReadingTime(notifications ? notifications : this.notifications)
@@ -91,13 +122,20 @@ export class NotificationsDropdownMenuComponent implements OnInit {
   }
 
   public deleteNotitification(notification: UserNotification) {
-    this.appService.deleteNotification(notification)
-      .subscribe((data) => {
-        this.logger.debug("Notification deleted", notification);
-        this.updateNotifications(
-          this.notifications.filter(n => n !== notification)
-        )
-      });
+    if (!notification) return;
+    if (notification.event === 'UNCONFIGURED_EMAIL')
+      this.updateNotifications(
+        this.notifications.filter(n => n !== notification)
+      )
+    else {
+      this.appService.deleteNotification(notification)
+        .subscribe((data) => {
+          this.logger.debug("Notification deleted", notification);
+          this.updateNotifications(
+            this.notifications.filter(n => n !== notification)
+          )
+        });
+    }
   }
 
   public deleteAllNotifications() {
@@ -108,7 +146,6 @@ export class NotificationsDropdownMenuComponent implements OnInit {
         this.hideDropdownMenu();
       });
   }
-
 
   public navigateTo(url: string, params: object) {
     return this.router.navigate([url, params]);
